Add 'hlpc' to supercede 'dfhl'. Remove 'dfhl'.
A silly utility to filter and highlight text represented as percentages
with red or yellow colors using configurable warning and critical
thresholds and configurable header output.

I wrote this mostly as a rudimentary disk check used when logging in to
a server. This replaces 'dfhl' which was a little too primitive for my
desires.

This utility does not yet support configurable colors but could in the
future if I choose to care about that.